Fix git-init failure with git ssh creds 🔑
In case of a private git resources that uses `ssh` keys, `git`
wouldn't find the correct `.ssh/known_hosts`. This only happens in
case of the user of the container is `root`.

We now treat use `root` as a special case. It should now work for any
user — being root or not.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>