Adds "commented" website front matter to docs
This change adds the front matter needed for the website without altering the site markdown or adding titles. It does this by adding the front matter in comment tags.

The tutorial and install are not part of the website currently so links to them under docs/ use absolute urls to allow them to work in both the website and in the pipeline repo

(cherry picked from commit 00194731669ca5a279ca92b2d94cfb85f6ae7f6d)